requirement.txt
absl-py==1.4.0
appdirs==1.4.4
asttokens==2.2.1
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
backcall==0.2.0
batchgenerators==0.25
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
comm==0.2.2
connected-components-3d==3.16.0
contextlib2==21.6.0
contourpy==1.1.0
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
cytoolz @ file:///croot/cytoolz_1667465931118/work
dask @ file:///tmp/abs_994957d9-ec12-411f-b953-c010f9d489d10hj3gz4k/croots/recipe/dask-core_1658513209934/work
debugpy==1.6.6
decorator==5.1.1
edt==2.4.0
einops==0.6.1
executing==1.2.0
fastremap==1.14.1
filelock==3.12.0
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.42.1
fsspec @ file:///croot/fsspec_1670336592807/work
future==1.0.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grad-cam==1.5.0
grpcio==1.58.0
h5py==3.9.0
hid-parser==0.0.3
huggingface-hub==0.14.1
idna @ file:///croot/idna_1666125576474/work
imagecodecs @ file:///croot/imagecodecs_1677576717595/work
imageio==2.31.3
imops==0.8.8
importlib-metadata==6.0.0
importlib-resources==6.0.1
ipykernel==6.29.4
ipython==8.11.0
ipywidgets==8.1.2
jedi==0.18.2
joblib==1.3.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.3
linecache2==1.0.0
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
ml-collections==0.1.1
monai==1.1.0
nest-asyncio==1.5.6
networkx==3.1
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
nvidia-ml-py==12.535.133
nvitop==1.3.2
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.0
pandas==2.0.3
parso==0.8.3
partd @ file:///opt/conda/conda-bld/partd_1647245470509/work
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle @ file:///croot/pycodestyle_1674267221883/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.14.0
# Editable install with no version control (PYMIC==0.4.0.2)
-e /mnt/data1/ZhouFF/BraTS_selfsup/PyMIC-dev
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-git-info==0.8.3
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML @ file:///croot/pyyaml_1670514731622/work
pyzmq==25.0.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
SimpleCRF==0.2.1.1
SimpleITK==2.2.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data==0.6.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX @ file:///tmp/build/80754af9/tensorboardx_1621440489103/work
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.6.13
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz @ file:///croot/toolz_1667464077321/work
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.65.0
traceback2==1.4.0
traitlets==5.9.0
ttach==0.0.3
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
tzdata==2023.3
unittest2==1.1.0
urllib3 @ file:///croot/urllib3_1673575502006/work
vit-pytorch==1.2.2
wcwidth==0.2.6
Werkzeug==2.3.7
widgetsnbextension==4.0.10
zipp==3.15.0
